Bugfix: fix type of number for offset and size (#229)
* fix type of number for offset and size

* optimization for reading in deep speed framework.

* increase buffer to 16 MB

* revert changes

+++ dlio_benchmark/reader/indexed_binary_mmap_reader.py
@@ -57,10 +57,10 @@ class IndexedBinaryMMapReader(FormatReader):
             self.file_map_ibr[filename] = []
             bin_buffer_mmap = np.memmap(offset_file, mode='r', order='C')
             bin_buffer = memoryview(bin_buffer_mmap)
-            self.file_map_ibr[filename].append(np.frombuffer(bin_buffer, dtype=np.uint8))
+            self.file_map_ibr[filename].append(np.frombuffer(bin_buffer, dtype=np.uint64))
             bin_buffer_mmap = np.memmap(sz_file, mode='r', order='C')
             bin_buffer = memoryview(bin_buffer_mmap)
-            self.file_map_ibr[filename].append(np.frombuffer(bin_buffer, dtype=np.uint8))
+            self.file_map_ibr[filename].append(np.frombuffer(bin_buffer, dtype=np.uint64))
 
     @dlp.log
     def load_index(self):
